requirements.txt
Django>2.0
psycopg2-binary
djangorestframework
django-rest-framework-social-oauth2
pyJWT
cryptography
hyper
coreapi
